requirements.txt
absl-py==1.3.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.25.0
colorama==0.4.6
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
Django==4.1.3
django-cleanup==6.0.0
django-extensions==3.2.1
django-redis==5.2.0
django-sslserver==0.22
dlib==19.24.0
dnspython==2.3.0
et-xmlfile==1.1.0
eventlet==0.33.3
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==22.11.23
fonttools==4.38.0
greenlet==2.0.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib==3.6.2
mediapipe==0.9.0
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.36
protobuf==3.20.3
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
